restructure validation and serialiation content
